console logs removed

+++ services/staff.js
@@ -11,11 +11,11 @@ class StaffService {
             const { username, password } = params;
             const staffData = await staff.findOne({ username:username });
             if (!staffData) {
-                return { statusCode: 401, statusMessage: 'User not found', message: 'User not found' };
+                return { statusCode: 404, statusMessage: 'User not found' };
               }
               const isPasswordValid = await bcrypt.compare(password, staffData.password);
               if (!isPasswordValid) {
-                return { statusCode: 401, statusMessage: 'Incorrect password', message: 'Incorrect password' };
+                return { statusCode: 422, statusMessage: 'Incorrect password' };
               }
             if(isPasswordValid){
                let userData = {
@@ -26,15 +26,17 @@ class StaffService {
                return { statusCode: 200, statusMessage: 'Success', data: token };
             }
         } catch (err) {
-            console.log(err)
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
         }
       }
 
       async  getOrganizationById(params) {
         try {
-            const { organizationId } = params;
-            const data = await organization.findOne({organizationId:organizationId});
+            const { organizationId,userData } = params;
+            if (userData.roleId == 3) {
+              return { statusCode: 401, statusMessage: 'Unauthorized'  };
+            }
+            const data = await organization.findOne({organizationId:organizationId, isActive: true});
             return { statusCode: 200, statusMessage: 'Success', data: data };
         } catch (err) {
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
@@ -43,9 +45,11 @@ class StaffService {
 
       async  getAllOrganization(params) {
         try {
-          console.log("test--------------")
-            const data = await organization.find();
-            console.log("data",data)
+          const { userData } = params;
+          if (userData.roleId == 3) {
+            return { statusCode: 401, statusMessage: 'Unauthorized'  };
+          }
+            const data = await organization.find({ isActive: true });
             return { statusCode: 200, statusMessage: 'Success', data: data };
         } catch (err) {
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
@@ -54,13 +58,14 @@ class StaffService {
 
       async  deleteOrganization(params) {
         try {
-            const { organizationId } = params;
-            console.log("params",params)
+            const { organizationId,userData } = params;
+            if (userData.roleId == 3) {
+              return { statusCode: 401, statusMessage: 'Unauthorized'  };
+            }
             const data = await organization.updateOne(
                 { organizationId: organizationId, isActive: true },//condition
                 { $set: { isActive: false } } // Update
             );
-            console.log("data",data)
             return { statusCode: 200, statusMessage: 'Deleted Successfully', data: null };
         } catch (err) {
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
@@ -69,11 +74,14 @@ class StaffService {
 
       async  hardDeleteOrganization(params) {
         try {
-            const { organizationId } = params;
+            const { organizationId,userData } = params;
+            if (userData.roleId == 3) {
+              return { statusCode: 401, statusMessage: 'Unauthorized'  };
+            }
             const data = await organization.deleteOne(
                 { organizationId: organizationId}
             );
-            return { statusCode: 200, statusMessage: 'Deleted Successfully', data: data };
+            return { statusCode: 200, statusMessage: 'Deleted Successfully', data: null };
         } catch (err) {
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
         }
@@ -81,17 +89,26 @@ class StaffService {
 
       async  createOrganization(params) {
         try {
+          const { userData } = params;
+            if (userData.roleId == 3) {
+              return { statusCode: 401, statusMessage: 'Unauthorized'  };
+            }
             const organizationData = await organization.create(params);
            return { statusCode: 200, statusMessage: 'Success', data: organizationData };
         } catch (err) {
-            console.log(err)
+            if (err.code === 11000) {
+              return { statusCode: 409, statusMessage: 'organizationId should be unique' };
+            }
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
         }
       }
 
       async  editOrganization(params) {
-        try {
-            const { organizationId, description } = params;
+        try {            
+            const { organizationId, description,userData } = params;
+            if (userData.roleId == 3) {
+              return { statusCode: 401, statusMessage: 'Unauthorized'  };
+            }
             const data = await organization.updateOne(
                 { organizationId: organizationId, isActive: true },//condition
                 { description: description } // Update
@@ -108,7 +125,6 @@ class StaffService {
             const data = await user.findOne({userId:userId});  
             return { statusCode: 200, statusMessage: 'Success', data: data };
         } catch (err) {
-          console.log(err)
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
         }
       }
@@ -117,7 +133,6 @@ class StaffService {
         try {
             const { username, password } = params;
             const data = await user.findOne({ username:username });
-            console.log()
             if (!data) {
                 return { statusCode: 401, statusMessage: 'User not found', message: 'User not found' };
               }
@@ -134,7 +149,6 @@ class StaffService {
                return { statusCode: 200, statusMessage: 'Success', data: token };
             }
         } catch (err) {
-            console.log(err)
           return { statusCode: 500, statusMessage: 'Internal Server Error' };
         }
       }
